add spring boot spring data custom

+++ L9-spring-data-jpa-custom-starter/src/main/java/com/lss/l9springDataJpaCustom/methodExecutor/RepositoryMethodExecutor.java
@@ -1,33 +1,28 @@
-package com.lss.l9springDataJpaCustom.methodExtractor;
+package com.lss.l9springDataJpaCustom.methodExecutor;
 
-import com.lss.l9springDataJpaCustom.resolver.EntityNameResolver;
 import com.lss.l9springDataJpaCustom.query.JpqlQueryBuilder;
 import jakarta.persistence.EntityManager;
 import java.lang.reflect.Method;
 import java.util.List;
 
 public class RepositoryMethodExecutor {
-
     private final EntityManager em;
     private final Class<?> entityClass;
     private final String entityName;
 
-    public RepositoryMethodExecutor(EntityManager em, Class<?> entityClass) {
+    public RepositoryMethodExecutor(EntityManager em, Class<?> entityClass, String entityName) {
         this.em = em;
         this.entityClass = entityClass;
-        this.entityName = EntityNameResolver.resolve(entityClass);
+        this.entityName = entityName;
     }
 
     public Object execute(Method method, Object[] args) {
         String methodName = method.getName();
-
         return switch (methodName) {
             case "save" -> em.merge(args[0]);
             case "deleteById" -> {
                 Object entity = em.find(entityClass, args[0]);
-                if (entity != null) {
-                    em.remove(entity);
-                }
+                if (entity != null) em.remove(entity);
                 yield null;
             }
             case "findAll" -> em.createQuery("SELECT e FROM " + entityName + " e", entityClass).getResultList();
@@ -37,16 +32,11 @@ public class RepositoryMethodExecutor {
                     String jpql = JpqlQueryBuilder.buildFindByQuery(entityName, methodName);
                     var query = em.createQuery(jpql, entityClass);
                     String[] parts = methodName.substring(6).split("And");
-                    for (int i = 0; i < parts.length; i++) {
-                        query.setParameter("param" + i, args[i]);
-                    }
-                    if (method.getReturnType().isAssignableFrom(List.class)) {
-                        yield query.getResultList();
-                    } else {
-                        yield query.getSingleResult();
-                    }
+                    for (int i = 0; i < parts.length; i++) query.setParameter("param" + i, args[i]);
+                    if (method.getReturnType().isAssignableFrom(List.class)) yield query.getResultList();
+                    else yield query.getSingleResult();
                 }
-                throw new UnsupportedOperationException("Method not supported: " + methodName);
+                throw new UnsupportedOperationException("Unsupported method: " + methodName);
             }
         };
     }
